fix : statistics 폴더 네이밍 및 hirarchy 수정
