--- client/src/views/statistics/index.scss
@@ -1,111 +1,0 @@
-.content__statistics{
-    position:relative;
-    display:flex;
-    justify-content: space-between;
-    width:100%;
-    height:100%;
-}
-
-.expense-on-category{
-    flex:1;
-    margin-right:40px;
-    height:48%;
-    min-width:620px;
-    border-radius:10px;
-    background-color: #fff;
-    padding:40px 30px;
-    span {
-        font-family: Raleway;
-        font-weight: bold;
-        font-size: 30px;
-        color:var(--font-color);
-    }
-}
-
-.recently-account-history{
-    @extend .expense-on-category;
-    margin-right:0;
-    flex:1;
-
-    .recently-account-history__title{
-        display:block;
-        margin-bottom:63px;
-    }
-
-    .recently-account:not(:last-child){
-        margin-bottom:30px;
-    }
-}
-
-.recently-account-history__detail{
-    all:unset;
-    position:absolute;
-    right:35px;
-    top:50px;
-    cursor: pointer;
-}
-
-.curved-chart{
-    @extend .expense-on-category;
-    position:absolute;
-    bottom:0;
-    width:100%;
-}
-
-// RecentlyAccount
-.recently-account{
-    display:flex;
-    height: fit-content;
-    align-items:center;
-    padding: 0 20px;
-
-    img {
-        flex:1;
-        width:50px;
-    }
-    div{
-        flex:5;
-        margin-left:40px;
-        
-        span{
-            display:block;
-            text-align:start;
-            line-height:28px;
-            font-size:20px;
-            font-weight: normal;
-        }
-
-        .recently-account__title{
-            font-weight: bold;
-        }
-    }
-    span{
-        flex:3 160px;
-        font-size:25px;
-        text-align:end;
-        font-weight:bold;
-        .hundred-won{
-            font-weight: normal;
-        }
-    }
-}
-
-.expense-on-category__title{
-    display:block;
-}
-
-.expense-on-category__content{
-    display:flex;
-    height:280px;
-    svg {
-        flex:1;
-        // width:300px;
-        path.hide {
-            opacity:0;
-        }
-    }
-    .expense-on-category__sort{
-        flex:1;
-    }
-}
-
